feat: add framework-specific integration modules
