+++ src/ai_ml/anthropic_integration.py
@@ -1,0 +1,174 @@
+"""Anthropic Claude API integration for intelligent assistance."""
+
+from __future__ import annotations
+
+from datetime import datetime
+from typing import Any, Dict, List, Optional
+
+from .integration_common import get_env_var, logger
+
+try:
+    import anthropic
+except ImportError as exc:  # pragma: no cover - dependency not installed
+    raise ImportError(
+        "Anthropic package not available. Install with: pip install anthropic"
+    ) from exc
+
+
+class AnthropicIntegration:
+    """Wrapper around the Anthropic Claude client."""
+
+    def __init__(self, api_key: Optional[str] = None):
+        self.api_key = api_key or get_env_var("ANTHROPIC_API_KEY")
+
+        self.client = anthropic.Anthropic(api_key=self.api_key)
+        self.default_model = "claude-3-sonnet-20240229"
+        self.max_tokens = 4096
+        self.temperature = 0.7
+
+        logger.info("Anthropic integration initialized")
+
+    def generate_text(
+        self,
+        prompt: str,
+        model: Optional[str] = None,
+        max_tokens: Optional[int] = None,
+        temperature: Optional[float] = None,
+    ) -> str:
+        """Generate text using the Claude API."""
+        try:
+            message = self.client.messages.create(
+                model=model or self.default_model,
+                max_tokens=max_tokens or self.max_tokens,
+                temperature=temperature or self.temperature,
+                messages=[{"role": "user", "content": prompt}],
+            )
+
+            generated_text = message.content[0].text
+            logger.info("Generated text using %s", model or self.default_model)
+            return generated_text
+
+        except Exception as err:  # pragma: no cover - network errors
+            logger.error("Error generating text with Claude: %s", err)
+            return ""
+
+    def analyze_requirements(self, requirements_text: str) -> Dict[str, Any]:
+        """Analyze project requirements and provide insights."""
+        try:
+            prompt = f"""
+            Analyze the following project requirements and provide detailed insights:
+
+            {requirements_text}
+
+            Include:
+            1. Requirements completeness assessment
+            2. Potential technical challenges
+            3. Implementation recommendations
+            4. Risk assessment
+            5. Resource estimation
+            6. Alternative approaches
+            """
+
+            message = self.client.messages.create(
+                model=self.default_model,
+                max_tokens=self.max_tokens,
+                temperature=0.3,
+                messages=[{"role": "user", "content": prompt}],
+            )
+
+            analysis = message.content[0].text
+            logger.info("Requirements analysis completed")
+
+            return {
+                "analysis_type": "requirements_analysis",
+                "analysis": analysis,
+                "timestamp": datetime.now().isoformat(),
+                "model_used": self.default_model,
+            }
+
+        except Exception as err:  # pragma: no cover - network errors
+            logger.error("Error analyzing requirements with Claude: %s", err)
+            return {"error": str(err)}
+
+    def generate_documentation(self, code: str, doc_type: str = "README") -> str:
+        """Generate documentation for code."""
+        try:
+            prompt = f"""
+            Generate {doc_type} documentation for the following Python code:
+
+            {code}
+
+            Include:
+            1. Project overview
+            2. Installation instructions
+            3. Usage examples
+            4. API documentation
+            5. Configuration options
+            6. Troubleshooting guide
+            """
+
+            message = self.client.messages.create(
+                model=self.default_model,
+                max_tokens=self.max_tokens,
+                temperature=0.4,
+                messages=[{"role": "user", "content": prompt}],
+            )
+
+            documentation = message.content[0].text
+            logger.info("Generated %s documentation", doc_type)
+
+            return documentation
+
+        except Exception as err:  # pragma: no cover - network errors
+            logger.error("Error generating documentation with Claude: %s", err)
+            return ""
+
+    def code_review(
+        self, code: str, focus_areas: List[str] | None = None
+    ) -> Dict[str, Any]:
+        """Perform a comprehensive code review."""
+        try:
+            focus_areas = focus_areas or [
+                "quality",
+                "security",
+                "performance",
+                "maintainability",
+            ]
+
+            prompt = f"""
+            Perform a comprehensive code review of the following Python code:
+
+            {code}
+
+            Focus on: {', '.join(focus_areas)}
+
+            Provide:
+            1. Overall assessment (1-10 scale)
+            2. Specific issues found
+            3. Improvement suggestions
+            4. Security vulnerabilities
+            5. Performance optimizations
+            6. Best practices recommendations
+            """
+
+            message = self.client.messages.create(
+                model=self.default_model,
+                max_tokens=self.max_tokens,
+                temperature=0.3,
+                messages=[{"role": "user", "content": prompt}],
+            )
+
+            review = message.content[0].text
+            logger.info("Code review completed")
+
+            return {
+                "review_type": "comprehensive",
+                "focus_areas": focus_areas,
+                "review": review,
+                "timestamp": datetime.now().isoformat(),
+                "model_used": self.default_model,
+            }
+
+        except Exception as err:  # pragma: no cover - network errors
+            logger.error("Error performing code review with Claude: %s", err)
+            return {"error": str(err)}
